Move TestHelpers from iceberg-parquet to iceberg-api, use assertj assumptions
